+++ src/routes/adminRoutes.js
@@ -1,22 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const multer = require("multer");
-const path = require("path");
-
-//multer configuracion
-const storage = multer.diskStorage({
-  destination: (req, file, cb) => {
-    const destinationPath = path.join(__dirname, "../public/images/products");
-    cb(null, destinationPath);
-  },
-  filename: (req, file, cb) => {
-    const filename =
-      Date.now() + "-product-image" + path.extname(file.originalname);
-    cb(null, filename);
-  },
-});
-
-const upload = multer({ storage });
+const upload = require("../middlewares/multer");
 
 const adminController = require("../controllers/adminController");
 
@@ -26,8 +10,8 @@ router.post("/", upload.array("images"), adminController.create);
 
 router.get("/editproduct/:id", adminController.editProduct);
 
-router.put("/:id",upload.array("images"), adminController.edit);
+router.put("/:id", upload.array("images"), adminController.edit);
 
-router.delete("/:id",adminController.delete);
+router.delete("/:id", adminController.delete);
 
 module.exports = router;
